+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Image detection
 * ARKitLine
 * Custom light
+* Light estimation
 
 ## 0.0.3