+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.0.4
+## 0.1.0
 
 * Image detection
 * ARKitLine